pip_requirements.txt
absl-py==0.15.0
aiohttp==3.8.0
aiosignal==1.2.0
async-timeout==4.0.0
attrs==21.2.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
configparser==5.1.0
cycler==0.11.0
docker-pycreds==0.4.0
filelock==3.3.2
frozenlist==1.2.0
fsspec==2021.11.0
future==0.18.2
gitdb==4.0.9
GitPython==3.1.24
google-auth==2.3.3
google-auth-oauthlib==0.4.6
grpcio==1.41.1
higher==0.2.1
idna==3.3
jsonschema==4.2.1
kiwisolver==1.3.2
Markdown==3.3.4
matplotlib==3.4.3
msgpack==1.0.2
multidict==5.2.0
numpy==1.21.4
oauthlib==3.1.1
packaging==21.2
pandas==1.3.4
pathtools==0.1.2
Pillow==8.4.0
promise==2.3
protobuf==3.19.1
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.1
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytorch-lightning==1.5.0
pytz==2021.3
PyYAML==6.0
ray==1.8.0
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.7.2
sentry-sdk==1.4.3
shortuuid==1.0.3
six==1.16.0
smmap==5.0.0
subprocess32==3.5.4
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
torch==1.10.0+cu113
torchaudio==0.10.0+cu113
torchmetrics==0.6.0
torchvision==0.11.1+cu113
tqdm==4.62.3
typing-extensions==3.10.0.2
urllib3==1.26.7
wandb==0.12.7
Werkzeug==2.0.2
yarl==1.7.2
yaspin==2.1.0
